Se realizaron las siguientes modificaciones: - Correccion en la recoleccion y procesamiento de atributos de clase durante la preparacion de mapeo para DTOs y serializacion/deserializacion de JSON. - Mejoras significativas en generacion de expresiones SQL y su posterior conversion a cadenas. - Cambio significativo en el API expuesto par funcionalidades de mapeo de DTO, a manera de simplificar las responsabilidades de la clase y su utilizacion en tecnologias remotas (beans remotos). - Correccion de cambios en subpryecto de pruebas unitarias.

+++ Utils/src/main/java/fa/gs/utils/jsf/models/table/Pagination.java
@@ -13,22 +13,22 @@ import java.io.Serializable;
  */
 public class Pagination implements Serializable {
 
-    public static final Integer DEFAULT_LIMIT = 25;
+    public static final Long DEFAULT_LIMIT = 25L;
 
-    private final Integer limit;
-    private final Integer offset;
+    private final Long limit;
+    private final Long offset;
 
-    private Pagination(Integer limit, Integer offset) {
+    private Pagination(Long limit, Long offset) {
         this.limit = limit;
         this.offset = offset;
     }
 
-    public static Pagination fromPages(Integer page, Integer pageSize) {
+    public static Pagination fromPages(Long page, Long pageSize) {
         // Limite, puede ser nulo si no se define un offset.
-        Integer limit = pageSize;
+        Long limit = pageSize;
 
         // Offset, varia de acuerdo al valor del limite.
-        Integer offset = null;
+        Long offset = null;
         if (page != null) {
             if (limit == null) {
                 limit = DEFAULT_LIMIT;
@@ -39,16 +39,16 @@ public class Pagination implements Serializable {
         return fromLimits(limit, offset);
     }
 
-    public static Pagination fromLimits(Integer limit, Integer offset) {
+    public static Pagination fromLimits(Long limit, Long offset) {
         return new Pagination(limit, offset);
     }
 
     //<editor-fold defaultstate="collapsed" desc="Getters y Setters">
-    public Integer getLimit() {
+    public Long getLimit() {
         return limit;
     }
 
-    public Integer getOffset() {
+    public Long getOffset() {
         return offset;
     }
     //</editor-fold>
